lamnt98 commit chuyển Float thành Double và fix tạo và sửa xếp hạng 22/07

+++ project/src/main/java/com/example/webDemo3/service/impl/manageSchoolRankImpl/UpdateSchoolRankWeekServiceImpl.java
@@ -103,10 +103,10 @@ public class UpdateSchoolRankWeekServiceImpl implements UpdateSchoolRankWeekServ
         schoolRankWeekId.setWEEK_ID(responseDto.getWeekId());
         schoolRankWeekId.setSchoolClass(new Class(responseDto.getClassId()));
 
-        Float learningGrade = responseDto.getLearningGrade();
-        Float movementGrade = responseDto.getMovementGrade();
-        Float laborGrade = responseDto.getLaborGrade();
-        Float newTotalGrade = learningGrade + movementGrade + laborGrade + responseDto.getEmulationGrade();
+        Double learningGrade = responseDto.getLearningGrade();
+        Double movementGrade = responseDto.getMovementGrade();
+        Double laborGrade = responseDto.getLaborGrade();
+        Double newTotalGrade = learningGrade + movementGrade + laborGrade + responseDto.getEmulationGrade();
         schoolRankWeek.setSchoolRankWeekId(schoolRankWeekId);
         schoolRankWeek.setLearningGrade(learningGrade);
         schoolRankWeek.setMovementGrade(movementGrade);
